feat: minify output

+++ native-builder/utils/save-pages.go
@@ -1,16 +1,26 @@
 package utils
 
 import (
+	"bytes"
 	"fmt"
 	"garden-builder/types"
 	"os"
 	"path/filepath"
 	"text/template"
+
+	"github.com/tdewolff/minify/v2"
+	"github.com/tdewolff/minify/v2/css"
+	"github.com/tdewolff/minify/v2/html"
 )
 
 var layoutTemplate, layoutTemplateError = template.ParseFiles("views/layout")
+var minifier = loadHTMLMinifier()
 
 func SavePages(pages map[string]types.LayoutBuilderInput, directory string) {
+	if layoutTemplateError != nil {
+		panic(layoutTemplateError)
+	}
+
 	err := os.RemoveAll(directory)
 
 	if err != nil {
@@ -37,7 +47,10 @@ func savePage(path string, content types.LayoutBuilderInput, directory string) {
 		panic(err)
 	}
 
-	err = layoutTemplate.Execute(file, content)
+	rendered := new(bytes.Buffer)
+	err = layoutTemplate.Execute(rendered, content)
+
+	minifier.Minify("text/html", file, rendered)
 
 	if err != nil {
 		panic(err)
@@ -50,4 +63,12 @@ func savePage(path string, content types.LayoutBuilderInput, directory string) {
 	}
 
 	fmt.Println("✅ Saved page to", filePath)
+}
+
+func loadHTMLMinifier() *minify.M {
+	minifier := minify.New()
+	minifier.AddFunc("text/html", html.Minify)
+	minifier.AddFunc("text/css", css.Minify)
+
+	return minifier
 }
